Remove LegacyMockZipkinCollector and cleanup
