--- test/IntegrationTests/Helpers/LegacyMockZipkinCollector.cs
@@ -1,197 +1,0 @@
-// <copyright file="LegacyMockZipkinCollector.cs" company="OpenTelemetry Authors">
-// Copyright The OpenTelemetry Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-// </copyright>
-
-using System;
-using System.Collections.Generic;
-using System.Collections.Immutable;
-using System.Collections.Specialized;
-using System.IO;
-using System.Linq;
-using System.Net;
-using System.Text;
-using System.Threading.Tasks;
-using IntegrationTests.Helpers.Mocks;
-using IntegrationTests.Helpers.Models;
-using Newtonsoft.Json;
-using Xunit.Abstractions;
-
-namespace IntegrationTests.Helpers;
-
-public class LegacyMockZipkinCollector : IDisposable
-{
-    private static readonly TimeSpan DefaultSpanWaitTimeout = TimeSpan.FromMinutes(1);
-
-    private readonly object _syncRoot = new object();
-    private readonly ITestOutputHelper _output;
-    private readonly TestHttpListener _listener;
-
-    private LegacyMockZipkinCollector(ITestOutputHelper output, string host = "localhost")
-    {
-        _output = output;
-        _listener = new(output, HandleHttpRequests, host, "/api/v2/spans/");
-    }
-
-    public event EventHandler<EventArgs<HttpListenerContext>> RequestReceived;
-
-    public event EventHandler<EventArgs<IList<IMockSpan>>> RequestDeserialized;
-
-    /// <summary>
-    /// Gets or sets a value indicating whether to skip serialization of traces.
-    /// </summary>
-    public bool ShouldDeserializeTraces { get; set; } = true;
-
-    /// <summary>
-    /// Gets the TCP port that this collector is listening on.
-    /// </summary>
-    public int Port { get => _listener.Port; }
-
-    /// <summary>
-    /// Gets the filters used to filter out spans we don't want to look at for a test.
-    /// </summary>
-    public List<Func<IMockSpan, bool>> SpanFilters { get; private set; } = new List<Func<IMockSpan, bool>>();
-
-    private IImmutableList<IMockSpan> Spans { get; set; } = ImmutableList<IMockSpan>.Empty;
-
-    private IImmutableList<NameValueCollection> RequestHeaders { get; set; } = ImmutableList<NameValueCollection>.Empty;
-
-    public static async Task<LegacyMockZipkinCollector> Start(ITestOutputHelper output, string host = "localhost")
-    {
-        var collector = new LegacyMockZipkinCollector(output, host);
-
-        var healthzResult = await collector._listener.VerifyHealthzAsync();
-
-        if (!healthzResult)
-        {
-            collector.Dispose();
-            throw new InvalidOperationException($"Cannot start {nameof(MockLogsCollector)}!");
-        }
-
-        return collector;
-    }
-
-    /// <summary>
-    /// Wait for the given number of spans to appear.
-    /// </summary>
-    /// <param name="count">The expected number of spans.</param>
-    /// <param name="timeout">The timeout</param>
-    /// <param name="instrumentationLibrary">The integration we're testing</param>
-    /// <param name="minDateTime">Minimum time to check for spans from</param>
-    /// <param name="returnAllOperations">When true, returns every span regardless of operation name</param>
-    /// <returns>The list of spans.</returns>
-    public async Task<IImmutableList<IMockSpan>> WaitForSpansAsync(
-        int count,
-        TimeSpan? timeout = null,
-        string instrumentationLibrary = null,
-        DateTimeOffset? minDateTime = null,
-        bool returnAllOperations = false)
-    {
-        timeout ??= DefaultSpanWaitTimeout;
-        var deadline = DateTime.Now.Add(timeout.Value);
-        var minimumOffset = (minDateTime ?? DateTimeOffset.MinValue).ToUnixTimeNanoseconds();
-
-        IImmutableList<IMockSpan> relevantSpans = ImmutableList<IMockSpan>.Empty;
-
-        // Use a do-while to ensure at least one attempt at reading the data already
-        // received. This is helpful for negative tests, ie.: no spans generated, when
-        // the process emitting the spans already finished.
-        do
-        {
-            lock (_syncRoot)
-            {
-                relevantSpans =
-                    Spans
-                        .Where(s => SpanFilters.All(shouldReturn => shouldReturn(s)))
-                        .Where(s => s.Start > minimumOffset)
-                        .ToImmutableList();
-            }
-
-            if (relevantSpans.Count(s => instrumentationLibrary == null || s.Library == instrumentationLibrary) >= count)
-            {
-                break;
-            }
-
-            await Task.Delay(500);
-        }
-        while (DateTime.Now < deadline);
-
-        if (!returnAllOperations)
-        {
-            relevantSpans =
-                relevantSpans
-                    .Where(s => instrumentationLibrary == null || s.Library == instrumentationLibrary)
-                    .ToImmutableList();
-        }
-
-        return relevantSpans;
-    }
-
-    public void Dispose()
-    {
-        lock (_syncRoot)
-        {
-            WriteOutput($"Shutting down. Total spans received: '{Spans.Count}'");
-        }
-
-        _listener.Dispose();
-    }
-
-    protected virtual void OnRequestReceived(HttpListenerContext context)
-    {
-        RequestReceived?.Invoke(this, new EventArgs<HttpListenerContext>(context));
-    }
-
-    protected virtual void OnRequestDeserialized(IList<IMockSpan> trace)
-    {
-        RequestDeserialized?.Invoke(this, new EventArgs<IList<IMockSpan>>(trace));
-    }
-
-    private void HandleHttpRequests(HttpListenerContext ctx)
-    {
-        if (ShouldDeserializeTraces)
-        {
-            using (var reader = new StreamReader(ctx.Request.InputStream))
-            {
-                var zspans = JsonConvert.DeserializeObject<List<ZSpanMock>>(reader.ReadToEnd());
-                if (zspans != null)
-                {
-                    IList<IMockSpan> spans = zspans.ConvertAll(x => (IMockSpan)x);
-                    OnRequestDeserialized(spans);
-
-                    lock (_syncRoot)
-                    {
-                        Spans = Spans.AddRange(spans);
-                        RequestHeaders = RequestHeaders.Add(new NameValueCollection(ctx.Request.Headers));
-                    }
-                }
-            }
-        }
-
-        // NOTE: HttpStreamRequest doesn't support Transfer-Encoding: Chunked
-        // (Setting content-length avoids that)
-
-        ctx.Response.ContentType = "application/json";
-        var buffer = Encoding.UTF8.GetBytes("{}");
-        ctx.Response.ContentLength64 = buffer.LongLength;
-        ctx.Response.OutputStream.Write(buffer, 0, buffer.Length);
-        ctx.Response.Close();
-    }
-
-    private void WriteOutput(string msg)
-    {
-        const string name = nameof(LegacyMockZipkinCollector);
-        _output.WriteLine($"[{name}]: {msg}");
-    }
-}
