+++ test/IntegrationTests/Helpers/MockZipkinCollector.cs
@@ -17,16 +17,22 @@
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Net;
+using System.Runtime.Serialization;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
-using IntegrationTests.Helpers.Mocks;
 using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
 using Xunit;
 using Xunit.Abstractions;
+
+#if NETFRAMEWORK
+using IntegrationTests.Helpers.Compatibility;
+#endif
 
 namespace IntegrationTests.Helpers;
 
@@ -189,6 +195,143 @@ public class MockZipkinCollector : IDisposable
     {
         const string name = nameof(MockZipkinCollector);
         _output.WriteLine($"[{name}]: {msg}");
+    }
+
+    [DebuggerDisplay("TraceId={TraceId}, SpanId={SpanId}, Service={Service}, Name={Name}")]
+    public class ZSpanMock
+    {
+        [JsonExtensionData]
+        private Dictionary<string, JToken> _zipkinData;
+
+        public ZSpanMock()
+        {
+            _zipkinData = new Dictionary<string, JToken>();
+        }
+
+        public string TraceId
+        {
+            get => _zipkinData["traceId"].ToString();
+        }
+
+        public ulong SpanId
+        {
+            get => Convert.ToUInt64(_zipkinData["id"].ToString(), 16);
+        }
+
+        public string Name { get; set; }
+
+        public string Service
+        {
+            get => _zipkinData["localEndpoint"]["serviceName"].ToString();
+        }
+
+        public string Library { get; set; }
+
+        public ActivityKind Kind
+        {
+            get
+            {
+                if (_zipkinData.TryGetValue("kind", out var value))
+                {
+                    return (ActivityKind)Enum.Parse(typeof(ActivityKind), value.ToString(), true);
+                }
+
+                return ActivityKind.Internal;
+            }
+        }
+
+        public long Start
+        {
+            get => Convert.ToInt64(_zipkinData["timestamp"].ToString());
+        }
+
+        public long Duration { get; set; }
+
+        public ulong? ParentId
+        {
+            get
+            {
+                _zipkinData.TryGetValue("parentId", out JToken parentId);
+                return parentId == null ? null : Convert.ToUInt64(parentId.ToString(), 16);
+            }
+        }
+
+        public byte Error { get; set; }
+
+        public Dictionary<string, string> Tags { get; set; }
+
+        public Dictionary<DateTimeOffset, Dictionary<string, object>> Logs
+        {
+            get
+            {
+                var logs = new Dictionary<DateTimeOffset, Dictionary<string, object>>();
+
+                if (_zipkinData.TryGetValue("annotations", out JToken annotations))
+                {
+                    foreach (var item in annotations.ToObject<List<Dictionary<string, object>>>())
+                    {
+                        DateTimeOffset timestamp = ((long)item["timestamp"]).UnixMicrosecondsToDateTimeOffset();
+                        item.Remove("timestamp");
+                        logs[timestamp] = item;
+                    }
+                }
+
+                return logs;
+            }
+        }
+
+        public Dictionary<string, double> Metrics { get; set; }
+
+        public override string ToString()
+        {
+            var sb = new StringBuilder();
+            sb.AppendLine($"TraceId: {TraceId}");
+            sb.AppendLine($"ParentId: {ParentId}");
+            sb.AppendLine($"SpanId: {SpanId}");
+            sb.AppendLine($"Service: {Service}");
+            sb.AppendLine($"Name: {Name}");
+            sb.AppendLine($"Library: {Library}");
+            sb.AppendLine($"Kind: {Kind}");
+            sb.AppendLine($"Start: {Start}");
+            sb.AppendLine($"Duration: {Duration}");
+            sb.AppendLine($"Error: {Error}");
+            sb.AppendLine("Tags:");
+
+            if (Tags?.Count > 0)
+            {
+                foreach (var kv in Tags)
+                {
+                    sb.Append($"\t{kv.Key}:{kv.Value}\n");
+                }
+            }
+
+            sb.AppendLine("Logs:");
+            foreach (var e in Logs)
+            {
+                sb.Append($"\t{e.Key}:\n");
+                foreach (var kv in e.Value)
+                {
+                    sb.Append($"\t\t{kv.Key}:{kv.Value}\n");
+                }
+            }
+
+            return sb.ToString();
+        }
+
+        [OnDeserialized]
+        private void OnDeserialized(StreamingContext context)
+        {
+            Library = Tags.GetValueOrDefault("otel.library.name");
+
+            var error = Tags.GetValueOrDefault("error") ?? "false";
+            Error = (byte)(error.ToLowerInvariant().Equals("true") ? 1 : 0);
+
+            var spanKind = _zipkinData.GetValueOrDefault("kind")?.ToString();
+            if (spanKind != null)
+            {
+                Tags["span.kind"] = spanKind.ToLowerInvariant();
+            }
+        }
     }
 
     private class Expectation
