--- test/IntegrationTests/Helpers/Models/EventArgs.cs
@@ -1,29 +1,0 @@
-// <copyright file="EventArgs.cs" company="OpenTelemetry Authors">
-// Copyright The OpenTelemetry Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-// </copyright>
-
-using System;
-
-namespace IntegrationTests.Helpers.Models;
-
-public class EventArgs<T> : EventArgs
-{
-    public EventArgs(T value)
-    {
-        Value = value;
-    }
-
-    public T Value { get; }
-}
