+++ test/IntegrationTests/Helpers/TimeExtensions.cs
@@ -15,12 +15,15 @@
 // </copyright>
 
 using System;
-using IntegrationTests.Helpers.Constants;
 
 namespace IntegrationTests.Helpers;
 
 internal static class TimeExtensions
 {
+    private const long NanoSecondsPerTick = 1000000 / TimeSpan.TicksPerMillisecond;
+
+    private const long UnixEpochInTicks = 621355968000000000; // = DateTimeOffset.FromUnixTimeMilliseconds(0).Ticks
+
     /// <summary>
     /// Returns the number of nanoseconds that have elapsed since 1970-01-01T00:00:00.000Z.
     /// </summary>
@@ -28,7 +31,7 @@ internal static class TimeExtensions
     /// <returns>The number of nanoseconds that have elapsed since 1970-01-01T00:00:00.000Z.</returns>
     public static long ToUnixTimeNanoseconds(this DateTimeOffset dateTimeOffset)
     {
-        return (dateTimeOffset.Ticks - TimeConstants.UnixEpochInTicks) * TimeConstants.NanoSecondsPerTick;
+        return (dateTimeOffset.Ticks - UnixEpochInTicks) * NanoSecondsPerTick;
     }
 
     /// <summary>
@@ -40,7 +43,7 @@ internal static class TimeExtensions
     {
         const long microsecondsPerMillisecond = 1000;
         const long ticksPerMicrosecond = TimeSpan.TicksPerMillisecond / microsecondsPerMillisecond;
-        var ticks = (ts * ticksPerMicrosecond) + TimeConstants.UnixEpochInTicks;
+        var ticks = (ts * ticksPerMicrosecond) + UnixEpochInTicks;
         return new DateTimeOffset(ticks, TimeSpan.Zero);
     }
 }
